Remove "all" search, move search into own file
